General improvements
Cleaned up binning process and forced binning to occur even if not selected to get data into binned arrays for saving.

Upgraded saveAtoms to be able to save an arbitrary set of columnes from the schema, in any order chosen.

Started the process of splitting files into other modules to mimic the format of reductus (steps.py, loader.py).

Program probably does not compilie at this stage, but its time to quit for the day!